fix: isolate shared session membership and prune departed viewers
